[mvsData] Move ROI header from mvsUtils to mvsData
