+++ src/aliceVision/depthMap/Refine.hpp
@@ -7,9 +7,9 @@
 #pragma once
 
 #include <aliceVision/mvsData/StaticVector.hpp>
+#include <aliceVision/mvsData/ROI.hpp>
 #include <aliceVision/mvsUtils/MultiViewParams.hpp>
 #include <aliceVision/mvsUtils/ImagesCache.hpp>
-#include <aliceVision/mvsUtils/ROI.hpp>
 #include <aliceVision/depthMap/DepthSimMap.hpp>
 #include <aliceVision/depthMap/cuda/planeSweeping/similarity.hpp>
