+++ src/aliceVision/depthMap/cuda/planeSweeping/deviceDepthSimilarityMap.hpp
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <aliceVision/mvsUtils/ROI.hpp>
+#include <aliceVision/mvsData/ROI.hpp>
 #include <aliceVision/depthMap/RefineParams.hpp>
 #include <aliceVision/depthMap/cuda/host/DeviceCamera.hpp>
 #include <aliceVision/depthMap/cuda/host/memory.hpp>
