+++ src/aliceVision/depthMap/cuda/planeSweeping/deviceDepthSimilarityMapKernels.cuh
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <aliceVision/mvsUtils/ROI.hpp>
+#include <aliceVision/mvsData/ROI.hpp>
 #include <aliceVision/depthMap/cuda/device/buffer.cuh>
 #include <aliceVision/depthMap/cuda/device/matrix.cuh>
 #include <aliceVision/depthMap/cuda/device/Patch.cuh>
