+++ src/aliceVision/mvsData/ROI.hpp
