+++ src/aliceVision/mvsUtils/CMakeLists.txt
@@ -4,7 +4,6 @@ set(mvsUtils_files_headers
   fileIO.hpp
   ImagesCache.hpp
   MultiViewParams.hpp
-  ROI.hpp
   TileParams.hpp
 )
